[update]: useSlider shared 로 분리
