+++ packages/shared/src/index.ts
@@ -3,6 +3,7 @@ export * from './components';
 export * from './constants';
 export * from './dnd';
 export * from './examples';
+export * from './hooks';
 export * from './query';
 export * from './types';
 export * from './utils';
